Remove chapter parsing and move chunk_text_from_file to text_processing.py
